move binaryview to seperate files
